+++ taubsi/cogs/playerstats/link.py
@@ -1,8 +1,7 @@
 import discord
 
-from taubsi.taubsi_objects import tb
 from taubsi.cogs.setup.objects import TaubsiUser
-from taubsi.core.pogo import Team
+from taubsi.core import bot, Team
 
 
 class AcceptView(discord.ui.View):
@@ -10,11 +9,11 @@ class AcceptView(discord.ui.View):
         self.author_id = author_id
         super().__init__()
 
-    @discord.ui.button(label=tb.translate("link_send_again"), style=discord.ButtonStyle.blurple)
+    @discord.ui.button(label=bot.translate("link_send_again"), style=discord.ButtonStyle.blurple)
     async def send_code_again(self, _, interaction: discord.Interaction):
         if interaction.user.id != self.author_id:
             return
-        content = f"```\n{tb.config['botcode']}\n```"
+        content = f"```\n{bot.config['botcode']}\n```"
         await interaction.response.send_message(content=content,
                                                 ephemeral=True)
 
@@ -29,7 +28,7 @@ class LinkView(discord.ui.View):
         self.ingame = ingame
         self.interacted = False
 
-    @discord.ui.button(label=tb.translate("link_accept"), style=discord.ButtonStyle.green)
+    @discord.ui.button(label=bot.translate("link_accept"), style=discord.ButtonStyle.green)
     async def accept(self, _, interaction: discord.Interaction):
         if interaction.user.id != self.author.id or self.interacted:
             return
@@ -47,18 +46,18 @@ class LinkView(discord.ui.View):
             "user_id": self.author.id,
             "ingame_name": name
         }
-        await tb.intern_queries.insert("users", keyvals)
+        await bot.taubsi_db.insert("users", keyvals)
 
-        embed = discord.Embed(description=tb.translate("link_success").format(self.ingame[0][0]),
+        embed = discord.Embed(description=bot.translate("link_success").format(self.ingame[0][0]),
                               color=3092790)
         await self.message.edit(embed=embed, view=AcceptView(self.author.id))
         await interaction.response.send_message(
-            tb.translate("link_botcode").format(tb.config["botcode"]), ephemeral=True)
+            bot.translate("link_botcode").format(bot.config["botcode"]), ephemeral=True)
 
-    @discord.ui.button(label=tb.translate("link_deny"), style=discord.ButtonStyle.red)
+    @discord.ui.button(label=bot.translate("link_deny"), style=discord.ButtonStyle.red)
     async def deny(self, _, interaction: discord.Interaction):
         if interaction.user.id != self.author.id or self.interacted:
             return
         self.interacted = True
-        embed = discord.Embed(description=tb.translate("link_denied"), color=3092790)
+        embed = discord.Embed(description=bot.translate("link_denied"), color=3092790)
         await interaction.response.edit_message(embed=embed, view=None)
