+++ taubsi/cogs/playerstats/truexp.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
-from taubsi.cogs.playerstats.objects import Player, Stat, Badge
+
+from taubsi.cogs.playerstats.objects import Player
+from taubsi.core import Stat, Badge
 
 
 class TrueXP:
