+++ taubsi/core/__init__.py
@@ -1,3 +1,6 @@
 from .bot import TaubsiBot
-from .pogo import Gym, Raid
+from .pogo import Gym, Raid, Team, Stat, Badge
 from .logging import log
+
+
+bot = TaubsiBot()
